+++ src/main/java/roomescape/exception/client/BusinessRuleViolationException.java
@@ -1,6 +1,7 @@
-package roomescape.exception;
+package roomescape.exception.client;
 
 import org.springframework.http.HttpStatus;
+import roomescape.exception.base.RoomeScapeClientException;
 
 public class BusinessRuleViolationException extends RoomeScapeClientException {
     public BusinessRuleViolationException(String message) {
