+++ src/main/java/roomescape/exception/client/InvalidCommandException.java
@@ -1,6 +1,7 @@
-package roomescape.exception;
+package roomescape.exception.client;
 
 import org.springframework.http.HttpStatus;
+import roomescape.exception.base.RoomeScapeClientException;
 
 public class InvalidCommandException extends RoomeScapeClientException {
     public InvalidCommandException(String message) {
