+++ src/main/java/roomescape/exception/client/ResourceNotFoundException.java
@@ -1,6 +1,7 @@
-package roomescape.exception;
+package roomescape.exception.client;
 
 import org.springframework.http.HttpStatus;
+import roomescape.exception.base.RoomeScapeClientException;
 
 public class ResourceNotFoundException extends RoomeScapeClientException {
     public ResourceNotFoundException(String message) {
