+++ src/main/java/roomescape/exception/server/DataInconsistencyException.java
@@ -1,6 +1,7 @@
-package roomescape.exception;
+package roomescape.exception.server;
 
 import org.springframework.http.HttpStatus;
+import roomescape.exception.base.RoomeScapeServerException;
 
 public class DataInconsistencyException extends RoomeScapeServerException {
     public DataInconsistencyException(String message) {
