+++ src/main/java/roomescape/service/ReservationService.java
@@ -7,9 +7,9 @@ import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
 import roomescape.domain.policy.ReservationPolicy;
-import roomescape.exception.BusinessRuleViolationException;
-import roomescape.exception.DataInconsistencyException;
-import roomescape.exception.ResourceNotFoundException;
+import roomescape.exception.client.BusinessRuleViolationException;
+import roomescape.exception.client.ResourceNotFoundException;
+import roomescape.exception.server.DataInconsistencyException;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
