+++ src/main/java/roomescape/service/dto/ReservationUpdateCommand.java
@@ -1,7 +1,7 @@
 package roomescape.service.dto;
 
 import java.time.LocalDate;
-import roomescape.exception.InvalidCommandException;
+import roomescape.exception.client.InvalidCommandException;
 
 public class ReservationUpdateCommand {
     private final Long id;
