+++ src/test/java/roomescape/ReservationPolicyStepTest.java
@@ -16,8 +16,8 @@ import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Primary;
 import roomescape.domain.policy.ReservationPolicy;
-import roomescape.exception.BusinessRuleViolationException;
-import roomescape.exception.InvalidCommandException;
+import roomescape.exception.client.BusinessRuleViolationException;
+import roomescape.exception.client.InvalidCommandException;
 import roomescape.service.ReservationService;
 import roomescape.service.ReservationTimeService;
 import roomescape.service.dto.ReservationCreateCommand;
